requirements.txt
streamlit==1.25.0
streamlit-drawable-canvas==0.9.2
Pillow==10.0.0
numpy==1.24.2
